+++ sources/c/modules/app/include/low/itc/high/Semaphore.h
@@ -1,6 +1,6 @@
 typedef struct Semaphore {
     int (*wait)(struct Semaphore* self, int count);
-    int (*timewait)(struct Semaphore* self, int count, int timeout);
+    int (*timewait)(struct Semaphore* self, int count, long int timeout);
     int (*post)(struct Semaphore* self, int count);
     int (*get)(struct Semaphore* self);
 } Semaphore;
