+++ sources/c/modules/app/include/low/itc/low/Mutex.h
@@ -1,6 +1,6 @@
 typedef struct Mutex {
     int (*lock)(struct Mutex* self);
-    int (*timelock)(struct Mutex* self, int timeout);
+    int (*timelock)(struct Mutex* self, long int timeout);
     int (*unlock)(struct Mutex* self);
 } Mutex;
